HW1
The SalesTax.java file calculates the total price of products including their tax.
The Pay.java file calculates the wages that will be paid to workers who work less hours, regular hours and overtime. The code carters for all of the workers and ensures that everyone is paid according to the hours that they have worked.